requirements.txt
psycopg2==2.9.11
colorlog==6.9.0
PyYAML==6.0.3
paho-mqtt==2.1.0